Trying to solve benchmark settings logic. But a bit more refactoring is needed.

+++ ViewModel/SetupViewModel.cs
@@ -240,7 +240,7 @@ namespace GolemUI.ViewModel
             return true;
         }
 
-        public float? TotalHashRate => _benchmarkService.TotalMhs;
+        public double? TotalHashRate => _benchmarkService.TotalMhs;
 
         public double? ExpectedProfit
         {
